+++ DjangoApp/static/moj.css
@@ -8,7 +8,7 @@ body {
     background: url(../moje_static/images/background.jpg);
 }
 .header {
-    padding: 10px
+    padding: 10px;
     height: 6rem;
     text-align: left;
