+++ offerweb/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class OfferwebConfig(AppConfig):
     name = 'offerweb'
+    
+    def ready(self):
+        import offerweb.signals
