+++ offerweb/models.py
@@ -3,8 +3,7 @@ from phonenumber_field.modelfields import PhoneNumberField
 from address.models import AddressField
 from django.urls import reverse
 from .fields import OrderField
-from django.db.models.signals import post_save, pre_save
-from django.dispatch import receiver
+
 
 class Kontrahent(models.Model):
     Nazwa= models.CharField(max_length=32, unique=True, null = False)
@@ -34,11 +33,6 @@ class Oferty(models.Model):
     def numer_oferty(self):
             return "{} ({})".format(self.data, self.nr_zew)
     order = OrderField()
-@receiver ([post_save, pre_save], sender=Oferty)   
-def order_po_zapisaniu (sender, instance, **kwargs):
-     print("własnie stworzyłem numer oferty")    
-     print(instance.nr_zew)
-post_save.connect(order_po_zapisaniu,sender=Oferty) 
 
 class Indeksy(models.Model):
     mat={
